requirements.txt
feedparser
requests
pypdf
tiktoken
regex
ftfy
unidecode
tqdm
pymupdf
pdfplumber

flask
requests
flask-cors
python-dotenv
langchain
langchain-community
huggingface-hub
faiss-cpu
langchain-huggingface
sentence-transformers
